+++ MobileWallet/Pass.Manager/Pass.Manager.Web/Models/PassBeaconViewModel.cs
@@ -14,8 +14,8 @@ namespace Pass.Manager.Web.Models
         }
 
         public int PassBeaconId { get; set; }
-        public ushort? Major { get; set; }
-        public ushort? Minor { get; set; }
+        public int? Major { get; set; }
+        public int? Minor { get; set; }
         public string ProximityUuid { get; set; }
         public string RelevantText { get; set; }
